requirements.txt
pandas==2.2.2
numpy==1.26.4
xgboost==2.0.3
scikit-learn==1.4.2
netCDF4==1.6.5
matplotlib==3.8.4
streamlit==1.32.2
joblib==1.4.2
psycopg2-binary==2.9.9
openpyxl==3.1.5
folium==0.16.0
streamlit-folium==0.18.0
pyproj==3.6.1
